# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A complete REST API for managing multiple WhatsApp sessions using Go, Fiber, PostgreSQL, and whatsmeow library.\n\n## Authentication\nAll API endpoints (except /health and /swagger/*) require API key authentication.\nProvide your API key in the `Authorization` header.",
        "title": "zpwoot - WhatsApp Multi-Session API",
        "contact": {
            "name": "zpwoot Support",
            "url": "https://github.com/your-org/zpwoot",
            "email": "support@zpwoot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/sessions/{sessionId}/webhook/config": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the current webhook configuration for a specific session. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Get webhook configuration",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "\"123e4567-e89b-12d3-a456-426614174000\"",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook configuration retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/WebhookResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Session or webhook configuration not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new webhook configuration for a specific session. Webhooks will receive real-time events from Wameow. Requires API key authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Create webhook configuration",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "\"123e4567-e89b-12d3-a456-426614174000\"",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Webhook configuration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SetConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Webhook created successfully",
                        "schema": {
                            "$ref": "#/definitions/SetConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or parameters",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "SetConfigRequest": {
            "type": "object",
            "required": [
                "events",
                "url"
            ],
            "properties": {
                "events": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "message",
                        "status",
                        "connection"
                    ]
                },
                "secret": {
                    "type": "string",
                    "example": "my-webhook-secret-key-123"
                },
                "sessionId": {
                    "type": "string",
                    "example": "1b2e424c-a2a0-41a4-b992-15b7ec06b9bc"
                },
                "url": {
                    "type": "string",
                    "example": "https://myapp.com/webhook/whatsapp"
                }
            }
        },
        "SetConfigResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "message",
                        "status",
                        "connection"
                    ]
                },
                "id": {
                    "type": "string",
                    "example": "webhook-456def"
                },
                "sessionId": {
                    "type": "string",
                    "example": "1b2e424c-a2a0-41a4-b992-15b7ec06b9bc"
                },
                "url": {
                    "type": "string",
                    "example": "https://myapp.com/webhook/whatsapp"
                }
            }
        },
...
        "WebhookResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "message",
                        "status"
                    ]
                },
                "id": {
                    "type": "string",
                    "example": "webhook-123"
                },
                "sessionId": {
                    "type": "string",
                    "example": "session-123"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/webhook"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key directly (no Bearer prefix required). Example: dev-api-key-12345",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}